Rename component directories and move styling components to different dir
