+++ src/FiltersComponents/FilterDuration.tsx
